Remove fitbit_account model
The concept of a separate fitbit_account does not work as well now that
we're using omniauth to manage the login. We really only need the
oauth_token and oauth_secret from the login response, so we can store
those directly on the user record.

Along with that change we can simplify the way that the
FitgemClientWrapper is created.