Refactor the Backbone stack to use one-router-per-page
When the site was first created I made a single ResourcesRouter that handled routing for all of the resource types.
Over time I added more and more examples, and finally got to a point where it would be clearer if each resource
type used its own router since I just moved each resource type to its own page.

Included in this commit are also stubs for the foods router that will be used to create live data foods examples.