Save oauth credentials on every login
The Fitbit API changed recently to issue new user token/secret
credentials every time a user logs in, rather than to re-issue the same
credentials. That means that if we've stored previous oauth creds to the
database that we need to ensure the latest ones received overwrite the
previous ones.